chore: move util to plantir-extension
